Session

+++ server/controller/Cerrands.js
@@ -1,5 +1,5 @@
-const Errands = require('../models');
-const { Op } = require('sequelize');
+const Errands = require("../models");
+const { Op } = require("sequelize");
 // Access => Errands.User_info or Helper_board etc
 
 // ======= User sign =======
@@ -28,6 +28,7 @@ exports.userLogin = async (req, res) => {
     res.send(err);
   }
 };
+
 // ID 중복 검사
 exports.checkUserId = async (req, res) => {
   try {
@@ -35,10 +36,10 @@ exports.checkUserId = async (req, res) => {
       where: { user_id: { [Op.eq]: req.body.user_id } },
     });
     console.log(result);
-    if (result.dataValues.user_id) {
-      res.send(false);
+    if (!result) {
+      return res.send(false);
     } else {
-      res.send(true);
+      return res.send(true);
     }
   } catch (err) {
     res.send(err);
@@ -52,42 +53,10 @@ exports.checkUserName = async (req, res) => {
       where: { user_name: { [Op.eq]: req.body.user_name } },
     });
     if (!result) {
-      res.send(true);
+      return res.send(true);
     } else {
-      res.send(false);
+      return res.send(false);
     }
-  } catch (err) {
-    res.send(err);
-  }
-};
-
-// ID 중복 검사
-exports.checkUserId = async (req, res) => {
-  try {
-    const result = await Errands.User_info.findOne({
-      where: { user_id: { [Op.eq]: req.body.user_id } },
-    });
-    console.log(result);
-    if (!result) {
-      return res.send(true);
-    }
-    return res.send(false);
-  } catch (err) {
-    res.send(err);
-  }
-};
-
-// 닉네임 중복검사
-exports.checkUserName = async (req, res) => {
-  try {
-    const result = await Errands.User_info.findOne({
-      where: { user_name: { [Op.eq]: req.body.user_name } },
-    });
-    console.log(result);
-    if (!result) {
-      return res.send(true);
-    }
-    return res.send(false);
   } catch (err) {
     res.send(err);
   }
@@ -97,7 +66,7 @@ exports.checkUserName = async (req, res) => {
 exports.userRegister = async (req, res) => {
   try {
     const result = await Errands.User_info.findOne({
-      where: { user_id: req.body.user_id },
+      where: { user_id: { [Op.eq]: req.body.user_id } },
     });
     console.log(result);
     if (!result) {
@@ -114,13 +83,15 @@ exports.userRegister = async (req, res) => {
 };
 
 // 로그아웃
-exports.userLogout = async (req, res) => {
+exports.userLogout = (req, res) => {
   try {
-    await req.session.destroy((err) => {
+    console.log(req.session);
+    req.session.destroy((err) => {
       if (err) {
         throw err;
       }
-      req.session;
+      console.log("로그아웃 여부");
+      console.log(req.session);
       res.send(true);
     });
   } catch (err) {
@@ -177,14 +148,66 @@ exports.userLike = async (req, res) => {
   }
 };
 
+// 회원탈퇴
+exports.userWithdrawal = async (req, res) => {
+  try {
+    //   const auth = await Errands.User_info.findOne({
+    //      attributes: ['user_name'],
+    //      where: { user_name: { [Op.eq]: req.session.user_name } },
+    //    });
+    //    console.log(req.session.user_name);
+    //    if (auth.dataValues.user_name == req.session.user_name) {
+    const result = await Errands.User_info.destroy({
+      where: { user_id: { [Op.eq]: req.params.userId } },
+    });
+    console.log(result);
+    if (!result) {
+      res.send(false);
+    } else {
+      req.session.destroy();
+      res.send(true);
+    }
+  } catch (err) {
+    res.send(err);
+  }
+};
+// 회원정보 수정
+exports.update_user_info = async (req, res) => {
+  try {
+    // const auth = await Errands.User_info.findOne({
+    //   attributes: ['user_name'],
+    //   where: { user_name: { [Op.eq]: req.session.user_name } },
+    // });
+    // if (auth.dataValues.user_name == req.session.user_name) {
+    const [result] = await Errands.User_info.update(
+      {
+        user_id: req.body.user_id,
+        user_pw: req.body.user_pw,
+        user_name: req.body.user_name,
+        user_type: req.body.user_type,
+      },
+      { where: { user_id: { [Op.eq]: req.params.userId } } }
+    );
+    console.log(result);
+    if (result === 1) {
+      res.send(true);
+    } else {
+      res.send(false);
+    }
+  } catch (err) {
+    res.send(err);
+  }
+};
+
 // ======= Wanter_board =======
 // 매인페이지에 5개 보여주기 deadline순 5개
 exports.read_few_wanter_board = async (req, res) => {
   try {
     const result = await Errands.Wanter_board.findAll({
-      order: [['wanter_board_dead_line', 'desc']],
+      order: [["wanter_board_dead_line", "desc"]],
       limit: 5,
     });
+    console.log(req.session);
     res.send(result);
   } catch (err) {
     res.send(err);
@@ -195,7 +218,7 @@ exports.read_few_wanter_board = async (req, res) => {
 exports.read_wanter_board = async (req, res) => {
   try {
     const result = await Errands.Wanter_board.findAll({
-      order: [['wanter_board_date', 'asc']],
+      order: [["wanter_board_date", "asc"]],
     });
     res.send(result);
 
@@ -238,11 +261,11 @@ exports.create_wanter_board = async (req, res) => {
 exports.update_wanter_board = async (req, res) => {
   try {
     const auth = await Errands.User_info.findOne({
-      attributes: ['user_name'],
+      attributes: ["user_name"],
       where: { user_name: { [Op.eq]: req.session.user_name } },
     });
     console.log(auth);
-    if (auth.dataValues.user_name == req.session.user_name) {
+    if (auth) {
       const [result] = Errands.Wanter_board.update(
         {
           wanter_board_title: req.body.wanter_board_title,
@@ -257,7 +280,7 @@ exports.update_wanter_board = async (req, res) => {
       }
       res.send(true);
     } else {
-      res.send('작성자만 수정할 수 있습니다');
+      res.send("작성자만 수정할 수 있습니다");
     }
   } catch (err) {
     res.send(err);
@@ -268,10 +291,10 @@ exports.update_wanter_board = async (req, res) => {
 exports.delete_wanter_board = async (req, res) => {
   try {
     const auth = await Errands.User_info.findOne({
-      attributes: ['user_name'],
+      attributes: ["user_name"],
       where: { user_name: { [Op.eq]: req.session.user_name } },
     });
-    if (auth.dataValues.user_name == req.session.user_name) {
+    if (auth) {
       const result = Errands.Wanter_board.destroy({
         where: { wanter_board_id: { [Op.eq]: req.params.boardId } },
       });
@@ -280,7 +303,7 @@ exports.delete_wanter_board = async (req, res) => {
       }
       res.send(true);
     } else {
-      res.send('작성자만 삭제할 수 있습니다');
+      res.send("작성자만 삭제할 수 있습니다");
     }
   } catch (err) {
     res.send(err);
@@ -292,9 +315,7 @@ exports.delete_wanter_board = async (req, res) => {
 exports.read_wanter_comment = async (req, res) => {
   try {
     const result = await Errands.Wanter_comment.findAll({
-      where: {
-        wanter_comment_board_id: { [Op.eq]: req.params.boardId }
-      }
+      where: { wanter_comment_board_id: { [Op.eq]: req.params.boardId } },
     });
     res.send(result);
   } catch (err) {
@@ -320,10 +341,11 @@ exports.create_wanter_comment = async (req, res) => {
 exports.update_wanter_comment = async (req, res) => {
   try {
     const auth = await Errands.User_info.findOne({
-      attributes: ['user_name'],
+      attributes: ["user_name"],
+      // 수정필요
       where: { user_name: { [Op.eq]: req.session.user_name } },
     });
-    if (auth.dataValues.user_name == req.session.user_name) {
+    if (auth) {
       const [result] = Errands.Wanter_comment.update(
         {
           wanter_comment_content: req.body.wanter_comment_content,
@@ -340,7 +362,7 @@ exports.update_wanter_comment = async (req, res) => {
       }
       res.send(true);
     } else {
-      res.send('작성자만 삭제할 수 있습니다');
+      res.send("작성자만 수정할 수 있습니다");
     }
   } catch (err) {
     res.send(err);
@@ -353,10 +375,13 @@ exports.delete_wanter_comment = async (req, res) => {
     console.log(req.params);
     console.log(req.session.user_info);
     const auth = await Errands.User_info.findOne({
-      attributes: ['user_name'],
+      attributes: ["user_name"],
+      // 수정필요 req.params.commentId 테이블 찾고
       where: { user_name: { [Op.eq]: req.session.user_info.user_name } },
     });
     if (auth) {
+      console.log(auth);
+      console.log(req.session.user_info.user_name);
       const result = Errands.Wanter_comment.destroy({
         where: {
           wanter_comment_id: { [Op.eq]: req.params.commentId },
@@ -368,7 +393,7 @@ exports.delete_wanter_comment = async (req, res) => {
       }
       res.send(true);
     } else {
-      res.send('작성자만 댓글 삭제할 수 있습니다');
+      res.send("작성자만 댓글 삭제할 수 있습니다");
     }
   } catch (err) {
     res.send(err);
@@ -380,7 +405,7 @@ exports.delete_wanter_comment = async (req, res) => {
 exports.read_few_helper_board = async (req, res) => {
   try {
     const result = await Errands.Helper_board.findAll({
-      order: [['helper_board_date', 'asc']],
+      order: [["helper_board_date", "asc"]],
       limit: 3,
     });
     res.send(result);
@@ -437,10 +462,10 @@ exports.create_helper_board = async (req, res) => {
 exports.update_helper_board = async (req, res) => {
   try {
     const auth = await Errands.User_info.findOne({
-      attributes: ['user_name'],
+      attributes: ["user_name"],
       where: { user_name: { [Op.eq]: req.session.user_name } },
     });
-    if (auth.dataValues.user_name == req.session.user_name) {
+    if (auth) {
       const [result] = Errands.Helper_board.update(
         {
           helper_board_title: req.body.helper_board_title,
@@ -457,7 +482,7 @@ exports.update_helper_board = async (req, res) => {
       }
       res.send(true);
     } else {
-      res.send('작성자만 수정할 수 있습니다');
+      res.send("작성자만 수정할 수 있습니다");
     }
   } catch (err) {
     res.send(err);
@@ -468,7 +493,7 @@ exports.update_helper_board = async (req, res) => {
 exports.delete_helper_board = async (req, res) => {
   try {
     const auth = await Errands.User_info.findOne({
-      attributes: ['user_name'],
+      attributes: ["user_name"],
       where: { user_name: { [Op.eq]: req.session.user_name } },
     });
     if (auth.dataValues.user_name == req.session.user_name) {
@@ -480,7 +505,7 @@ exports.delete_helper_board = async (req, res) => {
       }
       res.send(true);
     } else {
-      res.send('작성자만 삭제할 수 있습니다');
+      res.send("작성자만 삭제할 수 있습니다");
     }
   } catch (err) {
     res.send(err);
@@ -492,7 +517,7 @@ exports.delete_helper_board = async (req, res) => {
 exports.read_helper_comment = async (req, res) => {
   try {
     const result = await Errands.Helper_comment.findAll({
-      wanter_comment_board_id: { [Op.eq]: req.params.boardId },
+      where: { helper_comment_board_id: { [Op.eq]: req.params.boardId } },
     });
     res.send(result);
   } catch (err) {
@@ -518,10 +543,10 @@ exports.create_helper_comment = async (req, res) => {
 exports.update_helper_comment = async (req, res) => {
   try {
     const auth = await Errands.User_info.findOne({
-      attributes: ['user_name'],
+      attributes: ["user_name"],
       where: { user_name: { [Op.eq]: req.session.user_name } },
     });
-    if (auth.dataValues.user_name == req.session.user_name) {
+    if (auth) {
       const [result] = Errands.Helper_comment.update(
         {
           helper_comment_content: req.body.helper_comment_content,
@@ -538,7 +563,7 @@ exports.update_helper_comment = async (req, res) => {
       }
       res.send(true);
     } else {
-      res.send('작성자만 수정할 수 있습니다');
+      res.send("작성자만 수정할 수 있습니다");
     }
   } catch (err) {
     res.send(err);
@@ -549,10 +574,10 @@ exports.update_helper_comment = async (req, res) => {
 exports.delete_helper_comment = async (req, res) => {
   try {
     const auth = await Errands.User_info.findOne({
-      attributes: ['user_name'],
+      attributes: ["user_name"],
       where: { user_name: { [Op.eq]: req.session.user_name } },
     });
-    if (auth.dataValues.user_name == req.session.user_name) {
+    if (auth) {
       const result = Errands.Helper_comment.destroy({
         where: {
           helper_comment_id: { [Op.eq]: req.params.commentId },
@@ -564,7 +589,7 @@ exports.delete_helper_comment = async (req, res) => {
       }
       res.send(true);
     } else {
-      res.send('작성자만 삭제할 수 있습니다');
+      res.send("작성자만 삭제할 수 있습니다");
     }
   } catch (err) {
     res.send(err);
@@ -592,6 +617,7 @@ exports.read_notice = async (req, res) => {
     const result = await Errands.Notice.findAll({
       order: [["wanter_board_date", "asc"]],
     });
+    res.send(result);
   } catch (err) {
     res.send(err);
   }
@@ -613,10 +639,10 @@ exports.read_one_notice = async (req, res) => {
 exports.create_notice = async (req, res) => {
   try {
     const auth = await Errands.User_info.findOne({
-      attributes: ['user_type'],
+      attributes: ["user_type"],
       where: { user_type: { [Op.eq]: req.session.user_type } },
     });
-    if (auth.dataValues.user_type == 'root') {
+    if (auth.dataValues.user_type == "root") {
       const result = Errands.Notice.create({
         notice_writer: req.session.user_name,
         notice_title: req.body.notice_title,
@@ -625,7 +651,7 @@ exports.create_notice = async (req, res) => {
 
       res.send(result);
     } else {
-      res.send('생성 권한이 없습니다');
+      res.send("생성 권한이 없습니다");
     }
   } catch (err) {
     res.send(err);
@@ -636,10 +662,10 @@ exports.create_notice = async (req, res) => {
 exports.update_notice = async (req, res) => {
   try {
     const auth = await Errands.User_info.findOne({
-      attributes: ['user_type'],
+      attributes: ["user_type"],
       where: { user_type: { [Op.eq]: req.session.user_type } },
     });
-    if (auth.dataValues.user_type == 'root') {
+    if (auth.dataValues.user_type == "root") {
       const [result] = Errands.Notice.update(
         {
           notice_title: req.body.notice_title,
@@ -652,7 +678,7 @@ exports.update_notice = async (req, res) => {
       }
       res.send(true);
     } else {
-      res.send('수정 권한이 없습니다');
+      res.send("수정 권한이 없습니다");
     }
   } catch (err) {
     res.send(err);
@@ -663,19 +689,22 @@ exports.update_notice = async (req, res) => {
 exports.delete_notice = async (req, res) => {
   try {
     const auth = await Errands.User_info.findOne({
-      attributes: ['user_type'],
+      attributes: ["user_type"],
       where: { user_type: { [Op.eq]: req.session.user_type } },
     });
-    if (auth.dataValues.user_type == 'root') {
-      const result = Errands.Notice.destroy({
-        where: { notice_id: { [Op.eq]: req.params.boardId } },
-      });
-      if (!result) {
-        return res.send(false);
+    console.log(auth);
+    if (auth) {
+      if (auth.dataValues.user_type == "root") {
+        const result = Errands.Notice.destroy({
+          where: { notice_id: { [Op.eq]: req.params.boardId } },
+        });
+        if (!result) {
+          return res.send(false);
+        }
+        return res.send(true);
       }
-      res.send(true);
     } else {
-      res.send('삭제 권한이 없습니다');
+      res.send("삭제 권한이 없습니다");
     }
   } catch (err) {
     res.send(err);
